The previous edit omitted an important restruction in Access Rules. (#2)
The edit to remove the SchemaEdit bit and simplify it to use a checkbox omitted the important restriction of other permissions that are granted by default to non-owners. This brings back that step.

Also adds a warning on top of the README in the hopes of attracting attention of anyone who might be affected.

Fix tests by force-installing an older chrome version that works with mocha-webdriver.